+++ src/Backend/Controlers/Typescript/Users/Users.ts
@@ -98,19 +98,19 @@ class User {
             const {Password,Name,Email}=req.body
             if(!Password||!Name||!Email){
                 return res.status(400).json({                    
-                    Message:"Parameters null",
-                    Status:400,
-                    Sucess:false
+                    "Message":"Parameters null",
+                    "Status":400,
+                    "Sucess":false
                 })
             }
             const response = await pool.query(
-                `SELECT id FROM "public"."Users" WHERE Email = $1;`,[Email]
+                `SELECT id FROM "public"."Users" WHERE "Email" = $1;`,[Email]
             );
             if (response.rows.length){
                 return res.status(400).json({
-                    Message:"Email already logged",
-                    Status:400,
-                    Sucess:false
+                    "Message":"Email already logged",
+                    "Status":400,
+                    "Sucess":false
                 })
             }
             else{
@@ -179,13 +179,21 @@ class User {
         }
     }
     private static async Update(id:number,User:User):Promise<MethodResponse>{
-        let response=await pool.query(`SELECT "Name","Email","Createdate" FROM "public"."Users" WHERE id = $1;`[id]);
+        let response=await pool.query(`SELECT "Password","Name","Email","Createdate" FROM "public"."Users" WHERE id = $1;`,[id]);
         if(!response.rows.length){
             return {
                 Message:"id not found",
                 Status:404,
                 Sucess:false
             }
+        }
+        const passwordMatch = await bcrypt.compare(User.Password, response.rows[0].Password);
+        if(!passwordMatch){
+            return {
+                Message:`Password Incorrect`,
+                Status:401,
+                Sucess:false
+            }    
         }
         else if(response.rows?.[0]){
             if(!User.Name){''
@@ -213,39 +221,60 @@ class User {
         try{
             const {id}= req.params
             const user=req.body
-            if(!user||!id){
+            if(!user||!id||!user.Password){
+                console.log(user,id)
                 return res.status(400).json({
                     Message:"Parameters null",
                     Status:400,
                     Sucess:false
                 })
             }
-            else if(typeof(id)!=="number"){
+            else if(isNaN(Number(id))){
                 return res.status(400).json({
                     Message:"Id not numeric",
                     Status:400,
                     Sucess:false
                 })
             }
-            let Retorno=await User.Update(id,user)
-            return res.status(400).json(Retorno)
+            let Retorno=await User.Update(parseInt(id, 10),user)
+            if(!Retorno.Sucess){
+                if(Retorno.Status==401){
+                    return res.status(401).json({
+                        Message:"Password Incorrect",
+                        Status:401,
+                        Sucess:false
+                    })
+                }
+                else if(Retorno.Status==404){
+                    return res.status(404).json({
+                        Message:"User not found",
+                        Status:404,
+                        Sucess:false
+                    })
+                }
+            }
+            return res.status(200).json({
+                Message:Retorno.Message,
+                Status:200,
+                Sucess:true
+            })
         }
         catch(e){
             console.error(e)
             return res.status(501).json(
                 {
                     Message:"Erro interno",
-                    Status:500,
+                    Status:501,
                     Sucess:false
                 }
             )
         }
     }
-    public static async Delete(id:number):Promise<MethodResponse>{
+    public static async Delete(id:number,Password:string):Promise<MethodResponse>{
         let response = await pool.query(
-            `SELECT id FROM "public"."Users" WHERE id = $1;`,
+            `SELECT "Password" FROM "public"."Users" WHERE id = $1;`,
             [id]
-        );        
+        );
         if(!response.rows.length){
             return {
                 Message:"User not found",
@@ -254,6 +283,14 @@ class User {
             }
         }
         try{
+            const passwordMatch = await bcrypt.compare(Password, response.rows[0].Password);
+            if(!passwordMatch){
+                return {
+                    Message:"Password Incorrect",
+                    Status:401,
+                    Sucess:false
+                }
+            }
             response=await pool.query(`DELETE FROM public."Users" WHERE id = $1;`,[id]);
             return {
                 Message:"Delete Successfully",
@@ -267,6 +304,7 @@ class User {
     }
     public static async DeleteRouter(req: Request, res: Response):Promise<Response>{
         const {id}= req.params
+        const {Password}= req.body
         if(!id){
             return res.status(400).json({
                 Message:"Parameters null",
@@ -282,10 +320,10 @@ class User {
                 Sucess:false
             })
         }
-        let response=await User.Delete(parseInt(id, 10))
+        let response=await User.Delete(parseInt(id, 10),Password)
         if(!response.Sucess){
             if(response.Status==404){
-                return res.status(400).json({
+                return res.status(404).json({
                     Message:"User not found",
                     Status:404,
                     Sucess:false
@@ -294,12 +332,21 @@ class User {
             else if(response.Status==501){
                 return res.status(501).json({
                     Message:"Internal error in database",
-                    Status:404,
+                    Status:501,
+                    Sucess:false
+                })
+            }
+            else if(response.Status==401){
+                return res.status(401).json({
+                    Message:"Password Incorrect",
+                    Status:401,
                     Sucess:false
                 })
             }
         }
+
         return res.status(200).json({
+            
             Message:"Deleted Successfully",
             Status:200,
             Sucess:true
